changes for ui style update
