+++ src/components/wishlist/common/Alert.css
@@ -1,0 +1,59 @@
+.swym-hl-alert-container{
+  position: fixed; 
+  right: 1rem; 
+  top: 9rem; 
+  z-index: 50; 
+  padding-top: 0.5rem;
+  padding-bottom: 0.5rem; 
+  padding-left: 1rem;
+  padding-right: 1rem; 
+  border-radius: 0.375rem; 
+  width: 20rem; 
+  text-align: left; 
+  color: #000000; 
+  color: #ffffff; 
+  background-color: #ffffff; 
+  transition-property: background-color, border-color, color, fill, stroke, opacity, box-shadow, transform;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  transition-duration: 300ms; 
+  box-shadow: 0 20px 25px -5px rgba(0, 0, 0, 0.1), 0 10px 10px -5px rgba(0, 0, 0, 0.04); 
+}
+
+.swym-hl-alert-content{
+  display: flex; 
+  margin-left: 0.5rem; 
+  align-items: center; 
+}
+
+.swym-hl-alert-img{
+  border-radius: 50%; 
+  border-width: 1px; 
+  width: 2.5rem; 
+}
+
+.swym-hl-alert-error-icon{
+  padding: 0.25rem; 
+  border-radius: 50%; 
+  color: #ffffff; 
+  background-color: #EF4444; 
+}
+
+.swym-hl-alert-success-icon{
+  padding: 0.25rem; 
+  border-radius: 50%; 
+  color: #ffffff; 
+  background-color: #44EF44;
+}
+
+.swym-hl-alet-info-container{
+  padding-left: 10px;
+}
+
+.swym-hl-alert-title{
+  color: #6B7280; 
+  font-weight: bold;
+}
+
+.swym-hl-alert-info{
+  color: #6B7280; 
+}
