+++ src/components/wishlist/index.js
@@ -1,3 +1,3 @@
-export {WishlistButton} from './WishlistButton.client';
-export {WishlistPage} from './WishlistPage.client';
-export {ShareWishlistPage} from './ShareWishlistPage.client';
+export {WishlistButton} from './wishlistButton/wishlistButton.client';
+export {WishlistPage} from './wishlistPage/WishlistPage.client';
+export {ShareWishlistPage} from './sharedList/ShareWishlistPage.client';
