+++ src/components/wishlist/shareListPopup/shareListPopup.css
@@ -1,0 +1,118 @@
+
+#swym-hl-share-list-popup{
+    position: fixed; 
+    top: 0;
+    right: 0;
+    bottom: 0;
+    left: 0; 
+    z-index: 10000; 
+    display: flex;
+    justify-content: center; 
+    align-items: center; 
+    width: 100vw; 
+    height: 100vh;
+    background-color: #00000061;
+}
+
+.swym-hl-share-modal-content {
+    margin: auto;
+    padding: 20px;
+    border: 1px solid #888;
+    display: flex; 
+    flex-direction: column; 
+    color: #000000; 
+    background-color: #ffffff; 
+    box-shadow: 0 20px 25px -5px rgba(0, 0, 0, 0.1), 0 10px 10px -5px rgba(0, 0, 0, 0.04); 
+    width: 600px;
+    position: relative;
+}
+  
+.swym-hl-share-modal-close-btn{
+    padding: 4px;
+    position: absolute;
+    top: 5px;
+    right: 15px;
+    font-size: 25px;
+    z-index: 100;
+    cursor: pointer;
+}
+
+.swym-hl-share-heading {
+    display: flex;
+    position: relative;
+    box-sizing: border-box;
+    border-bottom: 1px solid #828282;
+    padding-left: 0px;
+    padding-top: 0px;
+    padding-bottom: 10px;
+}
+
+.swym-hl-share-content {
+    margin-top: 20px;
+}
+
+.swym-hl-share-modal-loader{
+    display: flex;
+    justify-content: center;
+    align-item: center;
+    margin-right: 5px;
+}
+
+.swym-hl-share-modal-action-btn{
+    cursor: pointer;
+    padding: 8px 26px;
+    margin-top: 10px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    min-width: 170px;
+}
+
+.swym-hl-share-modal-action {
+    display: flex;
+    justify-content: space-between;
+    flex-wrap: wrap;
+}
+
+@media screen and (max-width: 450px) {
+    .swym-hl-share-modal-action-btn{
+        width: 100%;
+    }
+}
+
+.swym-input-label {
+    font-size: 12px;
+    line-height: 14px;
+    letter-spacing: 0.05em;
+    color: #828282;
+    margin-bottom: 5px;
+}
+
+.swym-share-wishlist-email,
+.swym-share-wishlist-name {
+    font-size: 14px;
+    letter-spacing: 0.05em;
+    width: 100%;
+}
+
+.swym-input-label{
+    padding-bottom: 20px;
+}
+
+.swym-input-has-error {
+    border: 1px solid #CC0000 !important;
+}
+
+.error-msg {
+    font-size: 12px;
+    line-height: 14px;
+    letter-spacing: 0.05em;
+    color: #CC0000;
+    font-style: italic;
+    padding: 5px;
+}
+
+.swym-title-new {
+    padding-top: 0px;
+    padding-left: 0px;
+}
