+++ src/components/wishlist/sharedList/ShareWishlistPage.css
@@ -1,0 +1,10 @@
+.swym-hl-shared-page-title{
+    font-size: 25px;
+    font-weight: bold;
+}
+
+.swym-hl-shared-page-list-container {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 10px;
+}
