+++ src/components/wishlist/wishlistButton/wishlistButton.css
@@ -1,0 +1,34 @@
+.swym-hl-bg-color{
+   background-color:#035587;
+}
+
+.swym-hl-bg-outline{
+    border: 1px solid gray;
+}
+
+.swym-hl-text-color{
+    color: #ffffff;  
+}
+
+.swym-hl-wl-icon{
+    width: fit-content;
+    background: transparent;
+    color: pink;
+}
+
+.swym-hl-wl-btn{
+    cursor: pointer; 
+    padding: 0.75rem;
+    border-radius: 0.25rem; 
+    font-weight: 500;
+}
+
+.swym-hl-btn-center-align{
+    align-items: center;
+    display: flex;
+    justify-content: center;
+}
+
+.swym-hl-product-wishlisted{
+    color: red;
+}
