+++ src/components/wishlist/wishlistItem/wishlistItem.css
@@ -1,0 +1,71 @@
+.swym-hl-listitem{
+    width: 24%;
+    position: relative;
+}
+
+
+.swym-hl-wishlist-item-content{
+    display: flex; 
+    overflow: hidden; 
+    flex-direction: column; 
+    flex: 1 1 0%; 
+    justify-content: space-between; 
+    border-radius: 0.25rem; 
+    height: 100%;
+    padding: 0.5rem; 
+    border-width: 1px; 
+    box-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.1), 0 4px 6px -2px rgba(0, 0, 0, 0.05); 
+}
+
+.swym-hl-listitem-delete-btn{
+    position: absolute;
+    top: 5px;
+    right: 5px;
+    padding: 8px;
+    cursor: pointer;
+    color: gray;
+    background: white;
+    border-radius: 50%;
+    box-shadow: 0 0.5px 0.7px 0.5px lightgray;
+}
+
+.swym-hl-list-item-title{
+    font-size: 16px;
+    margin-top: 5px;
+    margin-bottom: 5px;
+    font-weight: bold;
+}
+
+.swym-hl-list-item-vendor{
+    font-size: 14px;
+    margin-bottom: 10px;
+}
+
+.swym-hl-list-item-price{
+    font-size: 14px;
+    margin-bottom: 15px;
+    font-weight: bold;
+}
+
+.swym-hl-addtocart-btn{
+    padding: 1rem; 
+    border-radius: 0.25rem; 
+    width: 100%; 
+}
+
+@media only screen and (min-width: 600px) and (max-width: 900px) {
+    .swym-hl-listitem{
+        width: 32%;
+    }
+}
+
+@media screen and (max-width: 600px) {
+    .swym-hl-listitem{
+        width: 48%;
+    }
+}
+
+.swym-hl-listitem:hover{
+    transform: scale(1.05);
+    transition: transform 0.1s ease-in-out;
+}
