--- src/components/wishlist/wishlistPage.css
@@ -1,113 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-
-.swym-list-style {
-    padding: 10px 20px;
-    padding-bottom: 0px;
-    cursor: pointer;
-}
-
-.dropdown {
-  position: relative;
-  display: inline-block;
-}
-
-.dropdown-content {
-  display: none;
-  position: absolute;
-  background-color: #f9f9f9;
-  min-width: 160px;
-  box-shadow: 0px 8px 16px 0px rgba(0,0,0,0.2);
-  padding: 12px 16px;
-  z-index: 1;
-  margin: 10px;
-  margin-left: 0px
-}
-
-.swym-wishlist-style-ul{
-  margin:20px 0;
-  list-style:none;
-}
-
-.swym-wishlist-style-li{
-  position:relative;
-  padding:15px 10px;
-  border-bottom:1px solid rgba(255,255,255,.1);
-  overflow:hidden;
-  margin-right: 40px;
-  cursor: pointer;
-}
-
-.swym-wishlist-style-li:before{
-  content:'';
-  position:absolute;
-  left:-100%;
-  top:0;
-  width:100%;
-  height:100%;
-  background:#434655;
-  z-index:-1;
-}
-
-.swym-wishlist-style-li:hover:before{
-  left:0;
-  border-right:5px solid #434655;
-  opacity: 0.5;
-}
-
-.swym-wishlist-style-li a.active {
-  background-color: yellow;
-}
-
-
- .swym-wishlist-badge {
-  width: 30px;
-  height: 30px;
-  border-radius: 30px;
-  display: inline-block;
-  overflow: hidden;
-  font-size: 12px;
-  line-height: 20px;
-  text-align: center;
-  text-transform: uppercase;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-right: 14px;
-  flex-shrink: 0;
-}
-
-.selected {
-  background-color:#035587;
-  font-weight: bold;
-  font-size: 16px;
-  
-}
-
-.add-button{
-  background-color:#5a31f4;
-}
-
-.swym-spinner{
-  width:40px;
-  height: 40px;
-  border: 2px solid black;
-  border-radius: 25%;
-  animation: rotator 1.4s linear infinite;
-}
-  
-@keyframes rotator {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(270deg);
-  }
-}
-
-.grid-item-view:hover{
-  transform: scale(1.05);
-  transition: transform 0.1s ease-in-out;
-}
