+++ src/components/wishlist/wishlistPage/wishlistPage.css
@@ -1,0 +1,232 @@
+.swym-hl-wishlist-page{
+
+}
+
+.swym-hl-wl-dropdown-container{
+  display: inline-block; 
+  position: relative; 
+  text-align: left; 
+}
+
+.swym-hl-wl-list-dropdown-btn{
+  display: inline-flex; 
+  padding-top: 0.5rem;
+  padding-bottom: 0.5rem; 
+  padding-left: 1rem;
+  padding-right: 1rem; 
+  justify-content: center; 
+  border-width: 1px; 
+  border-color: #D1D5DB; 
+  width: 100%; 
+  font-size: 0.875rem;
+  line-height: 1.25rem; 
+  font-weight: 500; 
+  color: #374151; 
+  background-color: #ffffff; 
+  box-shadow: 0 1px 2px 0 rgba(0, 0, 0, 0.05); 
+}
+
+.swym-hl-dropdown-icon{
+  margin-left: 0.5rem; 
+  margin-right: -0.25rem; 
+  width: 1.25rem; 
+  height: 1.25rem; 
+}
+
+.swym-hl-wl-dropdown-list{
+  position: absolute; 
+  left: 0; 
+  z-index: 10; 
+  margin-top: 0.5rem; 
+  //box-shadow: var(--tw-ring-inset) 0 0 0 calc(1px + var(--tw-ring-offset-width)) var(--tw-ring-color); 
+  --ring-color: #000000; 
+  --ring-opacity: 0.05; 
+  width: 14rem; 
+  background-color: #ffffff; 
+  transform-origin: top right; 
+  box-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.1), 0 4px 6px -2px rgba(0, 0, 0, 0.05); 
+  border: 1px solid lightgray;
+  border-radius: 4px;
+}
+
+.swym-hl-wl-dropdown-list-item{
+  display: block; 
+  padding-top: 0.5rem;
+  padding-bottom: 0.5rem; 
+  padding-left: 1rem;
+  padding-right: 1rem; 
+  font-size: 0.875rem;
+  line-height: 1.25rem; 
+  color: #374151; 
+}
+
+.swym-hl-wl-dropdown-list-item.swym-hl-dropdown-list-item-active{
+  color: #111827; 
+  background-color: #F3F4F6; 
+}
+
+.swym-hl-wl-dropdown-list-item:hover{
+  color: #111827; 
+  background-color: #F3F4F6; 
+}
+
+.swym-hl-wl-list-dropdown-btn:hover{
+  background-color: #F9FAFB; 
+}
+
+.swym-hl-wishlist-page-title{
+  font-size: 20px;
+  font-weight: bold;
+}
+
+.swym-hl-wl-content-header{
+  display: flex;
+}
+
+.swym-hl-wl-selected-lname{
+  flex: 1 1 0%; 
+  font-size: 1.25rem;
+  line-height: 1.75rem; 
+  font-weight: 700; 
+}
+
+.swym-hl-wishlist-page-share-btn{
+  padding: 10px; 
+  border-radius: 9999px; 
+  border-width: 1px; 
+  cursor: pointer; 
+}
+
+.swym-hl-wishlist-page-list-container {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 10px;
+}
+
+.swym-list-style {
+    padding: 10px 20px;
+    padding-bottom: 0px;
+    cursor: pointer;
+}
+
+.dropdown {
+  position: relative;
+  display: inline-block;
+}
+
+.dropdown-content {
+  display: none;
+  position: absolute;
+  background-color: #f9f9f9;
+  min-width: 160px;
+  box-shadow: 0px 8px 16px 0px rgba(0,0,0,0.2);
+  padding: 12px 16px;
+  z-index: 1;
+  margin: 10px;
+  margin-left: 0px
+}
+
+.swym-wishlist-style-ul{
+  margin:20px 0;
+  list-style:none;
+}
+
+.swym-wishlist-style-li{
+  position:relative;
+  padding:15px 10px;
+  border-bottom:1px solid rgba(255,255,255,.1);
+  overflow:hidden;
+  margin-right: 40px;
+  cursor: pointer;
+}
+
+.swym-wishlist-style-li:before{
+  content:'';
+  position:absolute;
+  left:-100%;
+  top:0;
+  width:100%;
+  height:100%;
+  background:#434655;
+  z-index:-1;
+}
+
+.swym-wishlist-style-li:hover:before{
+  left:0;
+  border-right:5px solid #434655;
+  opacity: 0.5;
+}
+
+.swym-wishlist-style-li a.active {
+  background-color: yellow;
+}
+
+
+ .swym-wishlist-badge {
+  width: 30px;
+  height: 30px;
+  border-radius: 30px;
+  display: inline-block;
+  overflow: hidden;
+  font-size: 12px;
+  line-height: 20px;
+  text-align: center;
+  text-transform: uppercase;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-right: 14px;
+  flex-shrink: 0;
+}
+
+.selected {
+  background-color:#035587;
+  font-weight: bold;
+  font-size: 16px;
+  
+}
+
+.add-button{
+  background-color:#5a31f4;
+}
+
+.swym-spinner{
+  width:40px;
+  height: 40px;
+  border: 2px solid black;
+  border-radius: 25%;
+  animation: rotator 1.4s linear infinite;
+}
+  
+@keyframes rotator {
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(270deg);
+  }
+}
+
+.grid-item-view:hover{
+  transform: scale(1.05);
+  transition: transform 0.1s ease-in-out;
+}
+
+.swym-hl-continue-shop-btn{
+  padding: 0.5rem 1rem;
+  margin-top: 1rem;
+  border-radius: 0.25rem;
+  font-weight: bold;
+}
+
+.swym-hl-continue-shop-btn:hover{
+  transform: scale(1.05);
+  transition: transform 0.1s ease-in-out;
+}
+
+
+.swym-hl-empty-wl-title{
+  font-size: 1.25rem;
+  line-height: 1.75rem; 
+  font-weight: 700; 
+}
