+++ src/lib/styleUtils.jsx
@@ -1,5 +1,5 @@
 export const INPUT_STYLE_CLASSES =
-  'appearance-none rounded dark:bg-transparent border focus:border-primary/50 focus:ring-0 w-full py-2 px-3 text-primary/90 placeholder:text-primary/50 leading-tight focus:shadow-outline';
+  'appearance-none rounded dark:bg-transparent border focus:border-primary/50 focus:ring-0 w-full py-2 px-3 leading-tight focus:shadow-outline';
 
 export const getInputStyleClasses = (isError) => {
   return `${INPUT_STYLE_CLASSES} ${
