+++ src/routes/products/[handle].server.jsx
@@ -88,7 +88,7 @@ export default function Product() {
               media={media.nodes}
               className="w-screen md:w-full lg:col-span-2"
             />
-            <div className="sticky md:-mb-nav md:top-nav md:-translate-y-nav md:h-screen md:pt-nav hiddenScroll md:overflow-y-scroll">
+            <div className="sticky md:-mb-nav md:top-nav md:h-screen hiddenScroll md:overflow-y-scroll">
               <section className="flex flex-col w-full max-w-xl gap-8 p-6 md:mx-auto md:max-w-sm md:px-0">
                 <div className="grid gap-2">
                   <Heading as="h1" format className="whitespace-normal">
@@ -143,6 +143,10 @@ const PRODUCT_QUERY = gql`
     product(handle: $handle) {
       id
       title
+      handle
+      featuredImage{
+        url
+      }
       vendor
       descriptionHtml
       media(first: 7) {
