+++ src/styles/index.css
@@ -167,10 +167,10 @@ shop-pay-button {
   }
 
   .card-image {
-    @apply relative rounded overflow-clip flex justify-center items-center;
+    @apply rounded overflow-clip flex justify-center items-center;
     &::before {
       content: ' ';
-      @apply z-10 absolute block top-0 left-0 w-full h-full shadow-border rounded;
+      @apply z-10 block top-0 left-0 w-full h-full shadow-border rounded;
     }
     img {
       @apply object-cover w-full aspect-[inherit];
