+++ src/swym/Utils/strings.js
